requirements.txt
alpaca-py>=0.30
yfinance>=0.2
pandas>=2.0
numpy>=1.26
scipy>=1.12
python-dotenv>=1.0
pytest>=8.0
